# openapi.yaml
openapi: 3.1.0
info:
  title: Persistant Agent API
  version: 0.1.0
paths:
  /:
    get:
      operationId: getServiceInfo
      responses:
        "200":
          description: Machine-readable service discovery and OpenAPI location.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInfoResponse"
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Agent process health.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /session:
    get:
      operationId: getSession
      responses:
        "200":
          description: Current managed Codex session state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
  /session/model:
    post:
      operationId: selectModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelSelectRequest"
      responses:
        "202":
          description: Model selection command was sent to Codex.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
  /messages:
    post:
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"
      responses:
        "202":
          description: Message was sent to Codex.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
  /context:
    post:
      operationId: updateContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextRequest"
      responses:
        "202":
          description: Context command was sent to Codex.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
components:
  schemas:
    ServiceInfoResponse:
      type: object
      required: [name, version, openapi_url, docs_url, commands]
      properties:
        name:
          type: string
        version:
          type: string
        openapi_url:
          type: string
        docs_url:
          type: string
        commands:
          type: object
          additionalProperties:
            type: string
    HealthResponse:
      type: object
      required: [status, codex_running]
      properties:
        status:
          type: string
          enum: [ok]
        codex_running:
          type: boolean
    SessionResponse:
      type: object
      required: [running, target_repo, command, pid]
      properties:
        running:
          type: boolean
        target_repo:
          type: string
        command:
          type: array
          items:
            type: string
        pid:
          type: integer
          nullable: true
    ModelSelectRequest:
      type: object
      required: [model]
      properties:
        model:
          type: string
          minLength: 1
        effort:
          type: string
          enum: [minimal, low, medium, high]
          nullable: true
    MessageRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
    ContextRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [empty, compact]
    CommandResponse:
      type: object
      required: [accepted, sent]
      properties:
        accepted:
          type: boolean
        sent:
          type: array
          items:
            type: string